Add conn, netAddr structs
Import conn struct from google.golang.com/test/bufconn. Extend this
implementation to simulate temporary disconnects and delays
(similar to the extension to pipe in the previous commit).

The net.Conn interface specifies methods that return address
information in form of the net.Addr interface. Add a netAddr struct
that implements this interface.